Update __init__.py
Added in new code that allows the passer to return none, which allows the Any Switch RGThree to see an attached node as "none" and therefore allows that switch to operate with a passer attached. This is to add function that lets us auto-switch between various options, eliminating user having to make manual choices in some templates.